requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
crypto==1.4.1
cryptography==2.8
Django==3.0.3
django-extensions==2.2.8
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-werkzeug-debugger-runserver==0.3.1
idna==2.9
mod-wsgi==4.7.1
Naked==0.1.31
Pillow==7.0.0
pycparser==2.19
pycrypto==2.6.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
shellescape==3.8.1
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
Werkzeug==1.0.0
